feat: Update GA ID, refine layout and SEO
Updates the Google Analytics ID to G-CXMPYNL337 and refactors the App and Layout components. The routing is adjusted to render child routes within the Layout component, improving structure.

The ClinicLogo component's aria-label and title are enhanced for better accessibility and clarity, explicitly mentioning the clinic's specialization and location.

SEO improvements include more specific meta descriptions and titles on the home page and a more dynamic title generation in the SEO component to ensure consistent branding across all pages.